Attest operational capture before workload imports
